Segmentation des composants
