+++ src/components/providers/SearchProvider.js
@@ -38,6 +38,7 @@ export default function SearchProvider(props) {
       fromLongitude: "",
       maxBikeKm: 50,
       minBikeKm: 3,
+      fromStation: null,
     })
   );
