+++ src/views/comparator/Results.js
@@ -9,7 +9,7 @@ import Transportation from "./results/Transportation";
 import gares from "../../gares.json";
 import { distance, point } from "@turf/turf";
 import Emoji from "components/base/Emoji";
-import getStation from "./wikidata";
+import { Stations } from "./Stations";
 
 const Wrapper = styled.div`
   flex: 1;
@@ -24,13 +24,14 @@ export default function Results() {
     carpool,
     uncertainty,
   } = useContext(TransportationContext);
-  const { itinerary } = useContext(SearchContext);
+  const { itinerary, setItinerary } = useContext(SearchContext);
 
   console.log("ITI", itinerary);
 
   const garesTo = garesProches(gares, itinerary, "to");
   const garesFrom = garesProches(gares, itinerary, "from");
   console.log(garesFrom, garesTo);
+
   if (!itinerary.fromLatitude || itinerary.fromLatitude === "") {
     return (
       <div>
@@ -58,7 +59,13 @@ export default function Results() {
     return (
       <div>
         <h3>📍 Les gares à proximité du départ</h3>
-        <Gares gares={garesFrom} count={6} />
+        <Stations
+          gares={garesFrom}
+          count={6}
+          onClick={(stationUIC) =>
+            setItinerary({ ...itinerary, fromStation: stationUIC })
+          }
+        />
       </div>
     );
   }
@@ -67,78 +74,13 @@ export default function Results() {
       <p>1️⃣ &nbsp;Voici les gares les plus proches</p>
       <h3>Départ</h3>
 
-      <Gares gares={garesFrom} />
+      <Stations gares={garesFrom} />
       <h3>Arrivée</h3>
-      <Gares gares={garesTo} count={20} />
+      <Stations gares={garesTo} count={20} />
       <p>2️⃣ &nbsp;La suite n'est pas encore implémentée :)</p>
     </Wrapper>
   );
 }
-
-const StationVignette = styled.li`
-  border: 4px solid ${(props) => props.theme.colors.second};
-  background: #ffffff9e;
-  margin: 0.6rem;
-  box-shadow: 0 1px 3px rgba(41, 117, 209, 0.12),
-    0 1px 2px rgba(41, 117, 209, 0.24);
-  padding: 0.6rem;
-  will-change: box-shadow;
-  -webkit-user-select: text;
-  user-select: text;
-  transition: box-shadow 0.15s, border-color 0.15s;
-  border-radius: 0.3rem;
-  .emoji {
-    font-size: 140%;
-  }
-  display: flex;
-  justify-content: space-between;
-`;
-
-const StationList = styled.ul`
-  list-style-type: none;
-`;
-
-const Station = ({ libelle, commune, distance, uic }) => {
-  const [data, setData] = useState(null);
-
-  useEffect(() => {
-    getStation(uic.slice(0, -1)).then((json) =>
-      setData(json?.results?.bindings[0])
-    );
-  }, [uic]);
-
-  return (
-    <StationVignette key={libelle}>
-      <div>
-        <strong>{libelle}</strong>
-        <div>
-          <Emoji>🚴</Emoji> {Math.round(distance)} km{" "}
-        </div>
-        <div>
-          {commune.toUpperCase() !== libelle.toUpperCase() && (
-            <span>
-              <Emoji>🏘️ </Emoji>&nbsp;
-              {commune}
-            </span>
-          )}
-        </div>
-      </div>
-      <div>{data?.pic && <StationImage src={data.pic.value} />}</div>
-    </StationVignette>
-  );
-};
-
-const StationImage = styled.img`
-  max-width: 8rem;
-`;
-
-const Gares = ({ gares, count = 3 }) => (
-  <StationList>
-    {gares.slice(0, count).map((station) => (
-      <Station {...station} />
-    ))}
-  </StationList>
-);
 
 const garesProches = (gares, itinerary, toOrFrom) =>
   gares
